Implement DeepL translation of Community Posts
When sending a Community Post notification, an
additional notification with no @mention is emitted
with the translated post content.

In order to make code DRY, a `fetchJson` utility was created.
Error handling is very granular!

To process the `MonadIO m => m (Either Text a)` return type of
`fetchJson`, custom operators `>>>=` and `<&>>=` have been created.
This is sure to make some readers scratch their heads but their
type signature + Haddock should be enough to grok them

Minor:
- Config.Type has been moved inside Config.BotConfig
- Config.BotConfig has been moved to BotConfig